Handle summons in a custom script
