+++ Data Files/Tamriel_Data.omwscripts
@@ -8,6 +8,6 @@ GLOBAL: scripts/TamrielData/global_magic.lua
 GLOBAL: scripts/TamrielData/global_miscspells.lua
 GLOBAL: scripts/TamrielData/global_summons.lua
 PLAYER, NPC, CREATURE: scripts/TamrielData/actor_magic.lua
-PLAYER, NPC, CREATURE: scripts/TamrielData/actor_summons.lua
 LOAD: scripts/TamrielData/load_magic.lua
+CUSTOM: scripts/TamrielData/actor_summons.lua
 CUSTOM: scripts/TamrielData/container_banish.lua
